Update index.ts
I did some more testing, and it seems the `E` generic you have here is necessary to do the proper inference of "as". This is mind bending stuff 😂 

But doing this we get the base typing of a component and the explicit related to polymorphic and the generic "as".

So without this I was having problems using other "as" values than what was set as the default.